pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sandbox-autoencoders"
version = "0.1.0"
description = "VAE experiments for ZX Spectrum longplay frames"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fsspec>=2024.12.0",
  "gradio>=5.20.0",
  "huggingface-hub>=1.6.0",
  "opencv-python-headless>=4.8.0",
  "Pillow>=10.0.0",
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "requests>=2.32.0",
  "s3fs>=2024.12.0",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "tqdm>=4.66.0",
  "wandb>=0.19.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
